AWS: Glue View Catalog implementation

+++ api/src/main/java/org/apache/iceberg/view/ViewHistoryEntry.java
@@ -34,5 +34,5 @@ public interface ViewHistoryEntry {
   /**
    * Returns ID of the new current version
    */
-  int versionId();
+  long versionId();
 }
